Struct libdecor::FrameRef

pub struct FrameRef(_);

A reference to a Frame used for decorating a Wayland surface.

Implementations

impl FrameRef

pub fn close(&self)

Close the window.

Roughly translates to [xdg_toplevel::Event::Close].

pub fn map(&self)

Map the window.

This will eventually result in the initial configure event.

pub fn is_floating(&self) -> bool

Return true if the window is floating.

A window is floating when it’s not maximized, tiled, fullscreen, or in any similar way with a fixed size and state. Note that this function uses the “applied” configuration. If this function is used in the ‘configure’ callback, the provided configuration has to be applied via commit first, before it will reflect the current window state from the provided configuration.

pub fn set_app_id<S: AsRef<str>>(&self, app_id: S)

Set the application ID of the window.

pub fn set_title<S: AsRef<str>>(&self, title: S)

Set the title of the window.

pub fn title(&self) -> String

Get the title of the window.

pub fn set_capabilities(&self, capabilities: Capabilities)

Set new capabilities of the window.

This determines whether e.g. a window decoration should show a maximize button, etc.

Setting a capability does not implicitly unset any other.

pub fn unset_capabilities(&self, capabilities: Capabilities)

Unset capabilities of the window.

The opposite of set_capabilities.

pub fn has_capability(&self, capabilities: Capabilities) -> bool

Check whether the window has any of the given capabilities.

pub fn show_window_menu(&self, seat: &WlSeat, serial: u32, x: i32, y: i32)

Show the window menu.

pub fn popup_grab<S: AsRef<str>>(&self, seat_name: S)

Issue a popup grab on the window. Call this when a [wayland_protocols::xdg_shell::client::xdg_popup::XdgPopup] is mapped, so that it can be properly dismissed by the decorations.

pub fn popup_ungrab<S: AsRef<str>>(&self, seat_name: S)

Release the popup grab. Call this when you unmap a popup.

pub fn translate_coordinate(&self, surface_x: i32, surface_y: i32) -> (i32, i32)

Translate content surface local coordinates to toplevel window local coordinates. This can be used to translate surface coordinates to coordinates useful for e.g. showing the window menu, or positioning a popup.

pub fn set_max_content_size(&self, width: i32, height: i32)

Set the max content size.

This translates roughly to [xdg_toplevel::XdgToplevel::set_max_size].

pub fn set_min_content_size(&self, width: i32, height: i32)

Set the min content size.

This translates roughly to [xdg_toplevel::XdgToplevel::set_min_size].

pub fn resize(&self, seat: &WlSeat, serial: u32, edge: ResizeEdge)

Initiate an interactive resize.

This roughly translates to [xdg_toplevel::XdgToplevel::resize].

pub fn _move(&self, seat: &WlSeat, serial: u32)

Initiate an interactive move.

This roughly translates to [xdg_toplevel::XdgToplevel::_move].

pub fn commit(&self, state: &State, configuration: Option<&Configuration>)

Commit a new window state. This can be called on application driven resizes when the window is floating, or in response to received configurations, i.e. from e.g. interactive resizes or state changes.

pub fn set_minimized(&self)

Minimize the window.

Roughly translates to [xdg_toplevel::XdgToplevel::set_minimized].

pub fn set_maximized(&self)

Maximize the window.

Roughly translates to [xdg_toplevel::XdgToplevel::set_maximized].

pub fn unset_maximized(&self)

Unmaximize the window.

Roughly translates to [xdg_toplevel::XdgToplevel::unset_maximized].

pub fn set_fullscreen(&self, output: Option<&WlOutput>)

Fullscreen the window.

Roughly translates to [xdg_toplevel::XdgToplevel::set_fullscreen].

pub fn unset_fullscreen(&self)

Unfullscreen the window.

Roughly translates to [xdg_toplevel::XdgToplevel::unset_fullscreen].

pub fn set_visibility(&self, visible: bool)

Set the visibility of the frame.

If an application wants to be borderless, it can set the frame visibility to false.

pub fn is_visible(&self) -> bool

Get the visibility of the frame.

pub fn set_parent(&self, parent: &FrameRef)

Set the parent of the window.

This can be used to stack multiple toplevel windows above or under each other.

pub fn xdg_surface(&self) -> Option<XdgSurface>

Get the associated [xdg_surface::XdgSurface] for content [wayland_client::protocol::wl_surface::WlSurface].

pub fn xdg_toplevel(&self) -> Option<XdgToplevel>

Get the associated [xdg_toplevel::XdgToplevel] for the content [wayland_client::protocol::wl_surface::WlSurface].